Add bar chart to trend identification page
