+++ src/loaders/comparison/index.ts
@@ -1,5 +1,51 @@
 import { fetcher } from "../../utils";
 
+export const horizontalBarLoader = async () => {
+  const casesData = await fetcher("epidemic/cases_malaysia.csv");
+
+  const quarterlyCases: Record<string, Record<string, number[]>> = {};
+
+  const cleanedData = casesData.filter((data: any) => {
+    return data.date && data.cases_new;
+  });
+
+  cleanedData.forEach((data: any) => {
+    const date = new Date(data.date);
+    const year = date.getFullYear();
+    const month = date.getMonth() + 1;
+    const quarter = Math.ceil(month / 3);
+    const quarterKey = `Q${quarter}`;
+
+    const cases = parseInt(data.cases_new, 10);
+
+    if (quarterlyCases[year]) {
+      if (quarterlyCases[year][quarterKey]) {
+        quarterlyCases[year][quarterKey].push(cases);
+      } else {
+        quarterlyCases[year][quarterKey] = [cases];
+      }
+    } else {
+      quarterlyCases[year] = { [quarterKey]: [cases] };
+    }
+  });
+
+  const horizontalBarData: { group: string; value: number; year: number }[] = [];
+
+  Object.entries(quarterlyCases).forEach(([year, quarterData]) => {
+    Object.entries(quarterData).forEach(([quarter, values]) => {
+      const sumValues = values.reduce((sum, current) => sum + current, 0);
+
+      horizontalBarData.push({
+        group: `${year} ${quarter}`,
+        value: sumValues,
+        year: parseInt(year, 10),
+      });
+    });
+  });
+
+  return horizontalBarData;
+};
+
 export const stackedBarLoader = async () => {
 	const death_state = await fetcher("epidemic/deaths_state.csv");
 	const case_state = await fetcher("epidemic/cases_state.csv");
